fix: add round tracker
