+++ src/components/games/botc/header/Header.tsx
@@ -4,16 +4,13 @@ import Typography from "@mui/material/Typography";
 import Snackbar from "@mui/material/Snackbar";
 import Alert from "@mui/material/Alert";
 import Grid from "@mui/material/Grid2";
-import Button from "@mui/material/Button";
-import ButtonGroup from "@mui/material/ButtonGroup";
 import { SelectChangeEvent } from "@mui/material";
 // Custom
 import EditPlayers from "./edit-players/EditPlayers";
-import InfoPopup from "../../common/info-popover/InfoPopup";
-import { BotCPlayer } from "../../../recoil/botc-atom";
-import { playerDist } from "../../../constants/botc";
-
-const numRounds = [0, 1, 2, 3, 4, 5, 6, 7];
+import InfoPopup from "../../../common/info-popover/InfoPopup";
+import { BotCPlayer } from "../../../../recoil/botc-atom";
+import { playerDist } from "../../../../constants/botc";
+import Tracker from "./Tracker";
 
 interface HeaderProps {
   botcPlayers: BotCPlayer[];
@@ -45,10 +42,6 @@ const Header = ({
   updateText,
 }: HeaderProps) => {
   const [hasToast, setHasToast] = React.useState(false);
-  const [round, setRound] = React.useState(0);
-  const [tracker, setTracker] = React.useState(
-    numRounds.map(() => botcPlayers.map(() => 0)),
-  );
 
   /** close the toast message */
   const handleClose = () => setHasToast(false);
@@ -57,18 +50,6 @@ const Header = ({
     newBotCGame();
     setHasToast(true);
   };
-
-  /** Handle tracker toggle */
-  const handleClick = (i: number) => () => {
-    const tempTracker = [...tracker[round]];
-    tempTracker[i] = (tracker[round][i] + 1) % 3;
-    const newTracker = [...tracker];
-    newTracker[round] = tempTracker;
-    setTracker(newTracker);
-  };
-
-  /** Update round */
-  const updateRound = (i: number) => () => setRound(i);
 
   return (
     <>
@@ -85,34 +66,7 @@ const Header = ({
         </Grid>
 
         <InfoPopup title="Tracker">
-          <Grid container spacing={1}>
-            {botcPlayers.map(({ name }, i) =>
-              i < numPlayers + numTravelers ? (
-                <Grid key={i} size={4}>
-                  <Button
-                    fullWidth
-                    variant={tracker[round][i] > 0 ? "contained" : "text"}
-                    color={tracker[round][i] !== 2 ? "primary" : "error"}
-                    onClick={handleClick(i)}
-                  >
-                    {name}
-                  </Button>
-                </Grid>
-              ) : null,
-            )}
-          </Grid>
-          <br aria-hidden />
-          <ButtonGroup fullWidth aria-label="Pick Round">
-            {numRounds.map((i) => (
-              <Button
-                key={i}
-                onClick={updateRound(i)}
-                variant={i === round ? "contained" : "outlined"}
-              >
-                {i + 1}
-              </Button>
-            ))}
-          </ButtonGroup>
+          <Tracker botcPlayers={botcPlayers} end={numPlayers + numTravelers} />
         </InfoPopup>
 
         <InfoPopup title="Players">
