+++ src/components/games/botc/header/edit-players/EditNameAndPos.tsx
