+++ src/components/games/botc/header/edit-players/ScriptControls.tsx
