+++ src/components/games/botc/header/edit-players/ScriptSelect.tsx
